* p4a-pips: (74 commits)
  dead code elimination introduced to replace, eventually, use_def_elimination
  Some more information for curious people + fix order of parameters for va_arg()
  some reformatting added; property ALIASING_ACROSS_FORMAL_PARAMETERS taken into account in references_may_conflict_p
  Make force_loop_fusion more pyps friendly
  update about transformers; paragraph added about va_arg field for Ronan
  Improved documentation and code cleaning in invariant code motion phase.
  Improved documentation of the RI.
  Improved developer guide on debugging.
  Still some distcheck fixes.
  Still some missing files for distcheck.
  Add missing file for distcheck.
  Some warning fixes + forgotten info for autopips
  adding a missing C intrinsic : difftime from time.h
  Fix a typo in complexity.
  Improve array_to_pointer.
  Some warning fixes.
  several typos fixed; I am surprised the Latex bug had not been detected by the amny daily recompilations
  add forgotten methods definitions for pointer_regions resources (thanks to Mehdi for spotting the problem).
  main change: handle instruction statements in in effects computations (mostly generated by controlizer)
  merge cobra modification into the trunk
  ...

+++ packages/PIPS/pips/src/Libs/sac/unroll.c
@@ -97,11 +97,11 @@ static void compute_variable_size(statement s, MinMaxVar* varwidth)
         varwidth->min = width;
 }
 
-static void simd_loop_unroll(statement loop_statement, int rate)
+static void simd_loop_unroll(statement loop_statement, intptr_t rate)
 {
   range r = loop_range(statement_loop(loop_statement));
   expression erange = range_to_expression(r,range_to_nbiter);
-  int irange;
+  intptr_t irange;
   if(expression_integer_value(erange,&irange) && irange <=rate) {
     full_loop_unroll(loop_statement);
   }
